requirements.txt
streamlit>=1.30.0
beautifulsoup4>=4.12.0
httpx>=0.26.0
anthropic>=0.18.0
openai>=1.10.0
ollama>=0.1.0
geopy>=2.4.0
pydantic>=2.5.0
pandas>=2.1.0
python-dotenv>=1.0.0
lxml>=5.0.0
# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
